+++ core/core/Cargo.toml
@@ -160,7 +160,6 @@ services-lakefs = []
 services-memcached = ["dep:fastpool"]
 services-memory = []
 services-mini-moka = ["dep:mini-moka"]
-services-moka = ["dep:moka"]
 services-mongodb = ["dep:mongodb", "dep:mongodb-internal-macros"]
 services-monoiofs = ["dep:monoio", "dep:flume"]
 services-mysql = ["dep:sqlx", "sqlx?/mysql"]
