+++ core/services/moka/src/lib.rs
@@ -15,28 +15,26 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::sync::Arc;
+#![cfg_attr(docsrs, feature(doc_cfg))]
+//! Moka service implementation for Apache OpenDAL.
+#![deny(missing_docs)]
 
-use super::core::MokaCore;
-use crate::raw::oio;
-use crate::raw::*;
-use crate::*;
+mod backend;
+mod config;
+mod core;
+mod deleter;
+mod lister;
+mod writer;
 
-pub struct MokaDeleter {
-    core: Arc<MokaCore>,
-    root: String,
-}
+pub use backend::MokaBuilder as Moka;
+pub use backend::MokaCacheBuilder;
+pub use config::MokaConfig;
+pub use core::MokaValue;
 
-impl MokaDeleter {
-    pub fn new(core: Arc<MokaCore>, root: String) -> Self {
-        Self { core, root }
-    }
-}
+/// Default scheme for moka service.
+pub const MOKA_SCHEME: &str = "moka";
 
-impl oio::OneShotDelete for MokaDeleter {
-    async fn delete_once(&self, path: String, _: OpDelete) -> Result<()> {
-        let p = build_abs_path(&self.root, &path);
-        self.core.delete(&p).await?;
-        Ok(())
-    }
+#[ctor::ctor]
+fn register_moka_service() {
+    opendal_core::DEFAULT_OPERATOR_REGISTRY.register::<Moka>(MOKA_SCHEME);
 }
